requirements.txt
Jinja2>=3.1.4
PyYAML>=6.0.1
markdown>=3.5.2
